Add compatibility for Lootr.
Lootr is unable to hook off of the tile entity's `setLootTable` as it
doesn't give it access to a valid world. Changing to use the static
method of `LockableLootTileEntity` is identical in all aspects to your
current method, but it can act as a hook for Lootr.

Lootr is a mod that replaces all vanilla chests with loot tables with a
custom chest (both with visuals and functionality): each player who
opens such a chest has a custom, instanced inventory generated which
only they see.

This allows for better multiplayer server experiences as it discourages
players from exploring new chunks in search of loot when they can visit
structures that have previously been visited and "looted" by someone
else.

Guaranteed no more finding someone else's junk in your loot chest!